refactor: split do script do lobby
